==> mvp/00_data_generation.ipynb <==
# Data generation through simulation of credit card transactions

## Data generation process

The simulation will consist of five main steps (follows this [book chapter](https://fraud-detection-handbook.github.io/fraud-detection-handbook/Chapter_3_GettingStarted/SimulatedDataset.html)):

1. **Generation of customer profiles:** Every customer is different in their spending habits. This will be simulated by defining some properties for each customer. The main properties will be their geographical location, their spending frequency, and their spending amounts. The customer properties will be represented as a table, referred to as the customer profile table.

2. **Generation of terminal profiles:** Terminal properties will simply consist of a geographical location. The terminal properties will be represented as a table, referred to as the terminal profile table.

3. **Association of customer profiles to terminals:** We will assume that customers only make transactions on terminals that are within a radius of of their geographical locations. This makes the simple assumption that a customer only makes transactions on terminals that are geographically close to their location. This step will consist of adding a feature ‘list_terminals’ to each customer profile, that contains the set of terminals that a customer can use.

4. **Generation of transactions:** The simulator will loop over the set of customer profiles, and generate transactions according to their properties (spending frequencies and amounts, and available terminals). This will result in a table of transactions.

5. **Generation of fraud scenarios:** This last step will label the transactions as legitimate or genuine. This will be done by following three different fraud scenarios.

## Setup

%load_ext autoreload
%autoreload 2

from pathlib import Path
from src.utils import load_config

config_path = Path.cwd() / "config.yaml"
config_path.exists()

config = load_config(config_path)

## Step 1: Generate customer profiles

from src.data.generator import generate_customer_profiles_table

customer_df = generate_customer_profiles_table(n_customers=5)
customer_df.head()

assert len(customer_df) == 5
assert all([col in customer_df.columns for col in ["customer_id", "loc_long_coord", "loc_lat_coord", "mean_amount", "std_amount", "mean_nb_tx_per_day"]])
assert all([long >= 0 and long <= 100 for long in customer_df.loc_long_coord.values])
assert all([lat >= 0 and lat <= 100 for lat in customer_df.loc_lat_coord.values])
assert all([mean >= 5 and mean <= 100 for mean in customer_df.mean_amount.values])
assert all([std >= 2.5 and std <= 50 for std in customer_df.std_amount.values])
assert all([mean >= 0 and mean <= 4 for mean in customer_df.mean_nb_tx_per_day.values])

## Step 2: Generate terminal profiles

from src.data.generator import generate_terminal_profiles_table

terminal_df = generate_terminal_profiles_table(n_terminals=5)
terminal_df.head()

assert len(terminal_df) == 5
assert all(col in terminal_df.columns for col in ["terminal_id", "loc_long_coord", "loc_lat_coord"])
assert all([long >= 0 and long <= 100 for long in terminal_df["loc_long_coord"]])
assert all([lat >= 0 and lat <= 100 for lat in terminal_df["loc_lat_coord"]])

## Step 3: Association of customer profiles to terminals

from src.data.generator import get_list_terminals_within_radius

# We first get the geographical locations of all terminals as a numpy array
x_y_terminals = terminal_df[['loc_long_coord','loc_lat_coord']].values.astype(float)
# And get the list of terminals within radius of 50 for the last customer
available_terminals = get_list_terminals_within_radius(customer_df.iloc[4], x_y_terminals=x_y_terminals, r=50)

assert len(available_terminals) == 2
assert all(term in available_terminals for term in [2, 3])

customer_df['available_terminals']=customer_df.apply(lambda x : get_list_terminals_within_radius(x, x_y_terminals=x_y_terminals, r=50), axis=1)

## Step 4: Generation of transactions

from src.data.generator import generate_transactions_table

transaction_df=generate_transactions_table(customer_df.iloc[0], 
                                                         start_date = "2018-04-01", 
                                                         nb_days = 5)
transaction_df.head()

assert all(col in transaction_df.columns for col in ["tx_time_seconds", "tx_time_days", "customer_id", "terminal_id", "tx_amount"])
assert all(amount >= 0 for amount in transaction_df.tx_amount.values)

transactions_df = customer_df.groupby('customer_id').apply(lambda x : generate_transactions_table(x.iloc[0], nb_days=5)).reset_index(drop=True)
transactions_df.head()

## Step 5: Generate a complete dataset

from src.data.generator import generate_dataset

(customer_profiles_table, terminal_profiles_table, transactions_df)=\
    generate_dataset(n_customers=5, 
                     n_terminals=5, 
                     nb_days=5, 
                     start_date="2018-04-01", 
                     r=50)

gen_config = config["data"]["generator"]

(customer_profiles_table, terminal_profiles_table, transactions_df)=\
    generate_dataset(n_customers=gen_config["num_customers"],
                     n_terminals=gen_config["num_terminals"],
                     nb_days=gen_config["num_days"],
                     start_date=gen_config["start_date"],
                     r=gen_config["customer_radius"])

## Step 6: Add fraudulent transactions

This last step of the simulation adds fraudulent transactions to the dataset, using the following fraud scenarios:

- **Scenario 1:** Any transaction whose amount is more than 220 is a fraud. This scenario is not inspired by a real-world scenario. Rather, it will provide an obvious fraud pattern that should be detected by any baseline fraud detector. This will be useful to validate the implementation of a fraud detection technique.

- **Scenario 2:** Every day, a list of two terminals is drawn at random. All transactions on these terminals in the next 28 days will be marked as fraudulent. This scenario simulates a criminal use of a terminal, through phishing for example. Detecting this scenario will be possible by adding features that keep track of the number of fraudulent transactions on the terminal. Since the terminal is only compromised for 28 days, additional strategies that involve concept drift will need to be designed to efficiently deal with this scenario.

- **Scenario 3:** Every day, a list of 3 customers is drawn at random. In the next 14 days, 1/3 of their transactions have their amounts multiplied by 5 and marked as fraudulent. This scenario simulates a card-not-present fraud where the credentials of a customer have been leaked. The customer continues to make transactions, and transactions of higher values are made by the fraudster who tries to maximize their gains. Detecting this scenario will require adding features that keep track of the spending habits of the customer. As for scenario 2, since the card is only temporarily compromised, additional strategies that involve concept drift should also be designed.

from src.data.generator import add_frauds

transactions_df = add_frauds(customer_profiles_table, terminal_profiles_table, transactions_df)

transactions_df.tx_fraud.mean()

transactions_df.tx_fraud.sum()

transactions_df.head()

The resulting dataset is interesting: It features class imbalance (less than 1% of fraudulent transactions), a mix of numerical and categorical features, non-trivial relationships between features, and time-dependent fraud scenarios.

## Step 7: Store data

DIR_OUTPUT = Path.cwd().parent / "data/raw/"
DIR_OUTPUT.mkdir(parents=True, exist_ok=True)

str(gen_config["start_date"])

import datetime

start_date = datetime.datetime.strptime(str(gen_config["start_date"]), "%Y-%m-%d")

for day in range(transactions_df.tx_time_days.max()+1):
    
    transactions_day = transactions_df[transactions_df.tx_time_days==day].sort_values('tx_time_seconds')
    
    date = start_date + datetime.timedelta(days=day)
    filename_output = date.strftime("%Y-%m-%d")+'.pkl'
    
    transactions_day.to_pickle(DIR_OUTPUT / filename_output)

customer_profiles_table.to_csv(DIR_OUTPUT / 'customer_profiles_table.csv', index=False)
terminal_profiles_table.to_csv(DIR_OUTPUT / 'terminal_profiles_table.csv', index=False)
