==> mvp/01_data_download.ipynb <==
# Data download

**Inputs:**

- Configuration file

**Steps:**

- Load configuration file
- Download data from source
- Transform into desired format
- Store data

**Outputs:**

- Raw data

## Setup

%load_ext autoreload
%autoreload 2

from pathlib import Path
from src.utils import load_config

config_path = Path.cwd() / "config.yaml"
config = load_config(config_path)

from dotenv import load_dotenv

load_dotenv()

data_dir = Path.cwd().parent / "data"
print(data_dir.exists())

## Step 1: Load and transform data

from src.data.io import load_dataframes

tx_df = load_dataframes(directory=data_dir / "raw", sort_by="tx_datetime")

print(tx_df.shape)

import pandas as pd

cust_df = pd.read_csv(data_dir / "raw/customer_profiles_table.csv")

print(cust_df.shape)

term_df = pd.read_csv(data_dir / "raw/terminal_profiles_table.csv")

print(term_df.shape)

proc_dir = data_dir / "processed"
proc_dir.mkdir(parents=True, exist_ok=True)

tx_df.to_csv(proc_dir / "transactions.csv", index=False)
cust_df.to_csv(proc_dir / "customers.csv", index=False)
term_df.to_csv(proc_dir / "terminals.csv", index=False)

## Step 3: Upload data to artifact store

import wandb

import os

wandb.login(key=os.environ["WANDB_API_KEY"], relogin=True, host="https://api.wandb.ai")

run = wandb.init(project="fraud-detection", job_type="data_download")

artifact = wandb.Artifact(name='raw_data', type='dataset')
artifact.add_dir(str(proc_dir))

run.log_artifact(artifact)

run.finish()
