==> mvp/02_exploratory_data_analysis.ipynb <==
# Exploratory data analysis

**Inputs:**

- Configuration file
- Raw data

**Steps:**

- Load configuration
- Load raw data
- Create summary statistics
- Create plots

**Outputs:**

- Summary statistics
- Plots

## Setup

%load_ext autoreload
%autoreload 2

from pathlib import Path
from src.utils import load_config

config_path = Path.cwd() / "config.yaml"
config = load_config(config_path)

from dotenv import load_dotenv

load_dotenv()

## Step 1: Load data

data_dir = Path.cwd().parent / "data/processed"
print(data_dir.exists())

import pandas as pd

customers = pd.read_csv(data_dir / "customers.csv")
terminals = pd.read_csv(data_dir / "terminals.csv")
transactions = pd.read_csv(data_dir / "transactions.csv")

print(customers.shape, terminals.shape, transactions.shape)

## Step 2: Perform exploratory data analysis

transactions.head()

### Create basic statistics

stats = {}

stats["number_of_transactions"] = len(transactions)
stats["number_of_customers"] = len(customers)
stats["number_of_terminals"] = len(terminals)
stats["start_date"] = transactions["tx_datetime"].min()
stats["end_date"] = transactions["tx_datetime"].max()

tx_per_day = transactions.groupby("tx_time_days")["transaction_id"].count()

# plot tx_per_day using plotly

import plotly.express as px

num_tx_fig = px.scatter(tx_per_day, x=tx_per_day.index, y="transaction_id")
num_tx_fig.update_xaxes(title_text="Day")
num_tx_fig.update_yaxes(title_text="Number of transactions")
num_tx_fig.add_hline(y=tx_per_day.mean(), line_dash="dot", annotation_text="Mean", annotation_position="bottom right")
num_tx_fig.add_hrect(y0=tx_per_day.mean()-tx_per_day.std(), y1=tx_per_day.mean()+tx_per_day.std(), line_width=0, fillcolor="red", opacity=0.2, annotation_text="Std", annotation_position="bottom right")
num_tx_fig.show()

### Create transaction amount statistics

tx_stats = transactions["tx_amount"].describe()
stats["tx_amount_mean"] = tx_stats["mean"]
stats["tx_amount_std"] = tx_stats["std"]
stats["tx_amount_min"] = tx_stats["min"]
stats["tx_amount_max"] = tx_stats["max"]
stats["tx_amount_median"] = tx_stats["50%"]

### Create transaction per customer statistics

tx_per_customer = transactions.groupby('customer_id')['transaction_id'].count().describe()
stats["customer_tx_min"] = tx_per_customer["min"]
stats["customer_tx_max"] = tx_per_customer["max"]
stats["customer_tx_mean"] = tx_per_customer["mean"]
stats["customer_tx_median"] = tx_per_customer["50%"]

### Create transaction per terminal statistics

tx_per_terminal = transactions.groupby('terminal_id')['transaction_id'].count().describe()
stats["terminal_tx_min"] = tx_per_terminal["min"]
stats["terminal_tx_max"] = tx_per_terminal["max"]
stats["terminal_tx_mean"] = tx_per_terminal["mean"]
stats["terminal_tx_median"] = tx_per_terminal["50%"]

### Create fraud statistics

frauds = transactions[transactions["tx_fraud"] == 1]
stats["percentage_of_frauds"] = len(frauds) / len(transactions)
stats["number_of_frauds"] = len(frauds)

frauds_by_scenario = frauds.groupby("tx_fraud_scenario")["transaction_id"].count()
stats["number_of_scenario1_frauds"] = frauds_by_scenario[1]
stats["number_of_scenario2_frauds"] = frauds_by_scenario[2]
stats["number_of_scenario3_frauds"] = frauds_by_scenario[3]

frauds_per_day = transactions.groupby("tx_time_days")["tx_fraud"].sum()

# Plot the number of frauds per day using plotly
frauds_per_day_fig = px.scatter(frauds_per_day, x=frauds_per_day.index, y=frauds_per_day.values, labels={"x": "Day", "y": "Number of frauds"})
frauds_per_day_fig.add_hline(y=frauds_per_day.mean(), line_dash="dot", annotation_text="Average", annotation_position="bottom right")
frauds_per_day_fig.show()

### Create fraud per customer statistics

frauds_per_customer = frauds.groupby('customer_id')['transaction_id'].count().describe()
stats["customer_fraud_min"] = frauds_per_customer["min"]
stats["customer_fraud_max"] = frauds_per_customer["max"]
stats["customer_fraud_mean"] = frauds_per_customer["mean"]
stats["customer_fraud_median"] = frauds_per_customer["50%"]

### Creat fraud per terminal statistics

frauds_per_terminal = frauds.groupby('terminal_id')['transaction_id'].count().describe()
stats["terminal_fraud_min"] = frauds_per_terminal["min"]
stats["terminal_fraud_max"] = frauds_per_terminal["max"]
stats["terminal_fraud_mean"] = frauds_per_terminal["mean"]
stats["terminal_fraud_median"] = frauds_per_terminal["50%"]

## Step 3: Log artifacts

stats_df = pd.DataFrame(data=[stats])
stats_df.head()

import wandb

run = wandb.init(project="fraud-detection", job_type="data_analysis")

run.use_artifact('felixpeters/fraud-detection/raw_data:v0', type='dataset')

eda_results = wandb.Artifact("eda_results", type="stats")

summary_stats = wandb.Table(dataframe=stats_df)

eda_results.add(summary_stats, "summary_stats")

img_dir = Path.cwd() / "images"
img_dir.mkdir(parents=True, exist_ok=True)

num_tx_fig_path = img_dir / 'num_tx_fig.png'
num_tx_fig.write_image(str(num_tx_fig_path))

num_tx_img = wandb.Image(str(num_tx_fig_path))

eda_results.add(num_tx_img, "number_of_transactions_per_day")

frauds_per_day_fig_path = img_dir / 'frauds_per_day_fig.png'
frauds_per_day_fig.write_image(str(frauds_per_day_fig_path))

frauds_per_day_img = wandb.Image(str(frauds_per_day_fig_path))

eda_results.add(frauds_per_day_img, "number_of_frauds_per_day")

run.log_artifact(eda_results)

run.finish()
