==> mvp/03_data_preprocessing.ipynb <==
# Data preprocessing

**Inputs:**

- Configuration file
- Raw data

**Steps:**

- Load configuration
- Load raw data
- Clean data
- Split data into training, validation, and test sets
- Engineer features
- Save data

**Outputs:**

- Training data
- Validation data
- Test data


## Setup

%load_ext autoreload
%autoreload 2

from pathlib import Path
from dotenv import load_dotenv
from src.utils import load_config

config_path = Path.cwd() / "config.yaml"
config = load_config(config_path)

load_dotenv()

## Step 1: Load data

import pandas as pd

data_dir = Path.cwd().parent / 'data'
proc_dir = data_dir / 'processed'

tx_df = pd.read_csv(proc_dir / "transactions.csv")
print(tx_df.shape)
cust_df = pd.read_csv(proc_dir / "customers.csv")
print(cust_df.shape)
term_df = pd.read_csv(proc_dir / "terminals.csv")
print(term_df.shape)

tx_df.tx_datetime = pd.to_datetime(tx_df.tx_datetime)

## Step 2: Create date and time features

from src.data.features import is_weekend

tx_df["tx_during_weekend"] = tx_df.tx_datetime.apply(is_weekend)

from src.data.features import is_night

tx_df["tx_during_night"] = tx_df.tx_datetime.apply(is_night)

## Step 3: Create customer features

from src.data.features import get_customer_spending_features

WINDOW_SIZES = config["data"]["features"]["window_sizes"]

tx_df = tx_df.groupby("customer_id").apply(lambda x: get_customer_spending_features(x, window_sizes=WINDOW_SIZES))

tx_df = tx_df.sort_values('tx_datetime').reset_index(drop=True)
tx_df.tail()

## Step 4: Create terminal features

from src.data.features import get_terminal_risk_features

DELAY_PERIOD = config["data"]["features"]["delay_period"]

tx_df = tx_df.groupby("terminal_id").apply(lambda x: get_terminal_risk_features(x, delay_period=DELAY_PERIOD, window_sizes=WINDOW_SIZES))
tx_df=tx_df.sort_values('tx_datetime').reset_index(drop=True)

tx_df.tail()

## Step 5: Store dataset

feature_dir = data_dir / "features"
feature_dir.mkdir(parents=True, exist_ok=True)

ds_path = feature_dir / "transactions.csv"

tx_df.to_csv(ds_path, index=False)

## Step 6: Track artifacts in MLOps platform

import wandb

run = wandb.init(project="fraud-detection", job_type="data_preprocessing")

run.use_artifact('felixpeters/fraud-detection/raw_data:v0', type='dataset')

features = wandb.Artifact("features", type="dataset")

features.add_dir(str(feature_dir))

run.log_artifact(features)

run.finish()
