==> mvp/04_model_training.ipynb <==
# Model training

**Inputs:**

- Configuration file
- Training data
- Validation data

**Steps:**

- Load configuration
- Load training data
- Load validation data
- Create model based on configuration
- Train model on training data
- Evaluate model on validation data

**Outputs:**

- Trained model
- Validation metrics

## Setup

%load_ext autoreload
%autoreload 2

from pathlib import Path
from dotenv import load_dotenv
from src.utils import load_config

config_path = Path.cwd() / "config.yaml"
config = load_config(config_path)

load_dotenv()

## Load historical data

data_dir = Path.cwd().parent / "data"
feature_dir = data_dir / "features"

assert data_dir.exists()
assert feature_dir.exists()

import pandas as pd

data = pd.read_csv(feature_dir / "transactions.csv")
data.tx_datetime = pd.to_datetime(data.tx_datetime)
print(data.shape)
data.head()

## Create train and test splits

import datetime

from src.data.split import get_train_test_set

train_start_date_str = config["data"]["split"]["start_date_training"]
delta_train = config["data"]["split"]["delta_train"]
delta_delay = config["data"]["split"]["delta_delay"]
delta_test = config["data"]["split"]["delta_test"]

train_start_date = datetime.datetime.strptime(str(train_start_date_str), "%Y-%m-%d")

train_df, test_df = get_train_test_set(data, train_start_date, delta_train=delta_train, delta_delay=delta_delay, delta_test=delta_test)

print("total transactions in training set:", len(train_df))
print("total transactions in test set:", len(test_df))
print("fraudulent transactions in training set:", len(train_df[train_df.tx_fraud == 1]))
print("fraudulent transactions in test set:", len(test_df[test_df.tx_fraud == 1]))

## Extract features

input_features = config["data"]["features"]["input_features"]
target_feature = config["data"]["features"]["output_feature"]

print(input_features)
print(target_feature)

## Train classifier

from src.model import fit_model
from sklearn.tree import DecisionTreeClassifier

classifier = DecisionTreeClassifier(max_depth=2, random_state=0)

results = fit_model(classifier, train_df, test_df, input_features, target_feature, scale=False)

test_df["tx_fraud_predicted"] = results["predictions_test"]
test_df.head()

## Calculate metrics

from src.metrics import evaluate_predictions

predictions_df = test_df
predictions_df["predictions"] = results["predictions_test"]

evals = evaluate_predictions(predictions_df, target_feature, "predictions", top_k_list=[100])
print(evals)

## Store artifacts

model_dir = data_dir / "models"
model_dir.mkdir(exist_ok=True, parents=True)

train_set_fp = model_dir / "train_set.csv"
test_set_fp = model_dir / "test_set.csv"
model_fp = model_dir / "model.joblib"

train_df.to_csv(train_set_fp, index=False)
test_df.to_csv(test_set_fp, index=False)

import joblib

joblib.dump(results["classifier"], model_fp)

## Track artifacts in MLOps platform

import wandb

run = wandb.init(project="fraud-detection", job_type="model_training")

run.use_artifact('felixpeters/fraud-detection/features:v0', type='dataset')

train_set_artifact = wandb.Artifact("training_set", type="dataset")
train_set_artifact.add_file(str(train_set_fp))
run.log_artifact(train_set_artifact)

test_set_artifact = wandb.Artifact("test_set", type="dataset")
test_set_artifact.add_file(str(test_set_fp))
run.log_artifact(test_set_artifact)

model_artifact = wandb.Artifact("model", type="model")
model_artifact.add_file(str(model_fp))
run.log_artifact(model_artifact)

run.log({
    "auc_roc": evals["auc_roc"],
    "average_precision": evals["average_precision"],
    "card_precision@100": evals["card_precision@100"],
})

run.finish()
